rename VariantGraph to VG
VG is more ambiguous, which is nice. This could be a "variation", a "variation
graph" or a "variant graph", or just a "VG". It's also easier to write!

This short name matches the other classes: Graph, Node, Edge, etc. It also
matches the library name, and the namespace name, but there aren't conflicts.

vg::VG ftw.