node trim
If we want to align to a portion of the graph where there are very long nodes,
we may know that we only want a certain part of the node. If we cut excess
sequence at the tail, then our alignment will still have correct positions
relative to the node start.  However, if we cut at the head of the node, then
our positions will not be sensible, and our alignments will have the wrong
start position. If we trim, we should record how much we trim from the 5' end
of the node. This provides a straightforward mechanism to transmit this
information.