correct label on critical region to avoid race
Use the graphq label when updating the completed set. This avoids a nasty race
condition that would leave several of the threads jumping through the main loop
even after the others had tried to exit. Judging by results in gdb this seems
to sometimes occur when a segfault happens.

Also update tests to be more robust (100 random trials for # of threads, region
size, and combinations of both).