allow a configurable maximum node size
Our ability to work with pieces of the graph is dependent on it being in
reasonable-sized chunks, not megabases, when we pick it up out of the index, or
work with it in parallel. It's a simple hack to get standard sizes. During construction, cut even regions of less than (or = to) the maximum node size between each successive variant.