merge graphs via a divide-and-conquer algorithm
Parallel construction runtime is dominated by the (ordered) reduce step. This
commit changes the reduce to an unordered one in which each thread seeks the
first case of a successive pair of completed graphs, combines these, and stores
the result in the ordered queue of graphs for further merging. The final result
is a single graph in the queue. This allows the merge to be shared among many
threads, which should improve performance during the construction of large
graphs from VCFs or elsewhere.