compress graphs which are serialized to disk by default
This really makes no difference from a usability standpoint.

A more clever method would be to compress in blocks, as is done with bgzip.
This would allow indexing and other fun stuff. This could be the wrong pattern
to allow for such behavior though, as ideally we work from full node indexes,
which may include positional indexes.