save lots of effort in kmer iteration
By only examining kmers derived from the node we are looking at, we can
avoid duplicating kmers. Because each node is processed by a single
thread, this eliminates inter-thread duplication (a big win) and allows
dynamic scheduling with single-node granularity. This in turn bounds the
worst-case job for a given thread, and should eliminate long stalls on
pathological regions of the graph.

This lets us speed up kmer indexing 10-fold-ish!