add edge crossing limit to kpaths/kmers generation
When going previous or next from a given node, count the number of edges
we cross and bail out if it exceeds a user-defined cap (which is
~infinity) by default. This provides a simple way to avoid indexing
regions which have very high complexity, and are likely to reduce the
specificity of mapping in addition to adding excess weight to the kmer
index. This is likely to be important for longer kmer sizes.