use ordered kmer clustering to target mapping ("match threads")
With short kmers we maintain sensitivity. However, nonspecific matching of
kmers will lead to spurious target clusters (disjoint subgraphs) to align to.
To mitigate this problem without expensive graph traversal, we apply a
heuristic "match thread" discovery algorithm. This checks the approximate
previous nodes and also previous positions in the read to see if there was a
previous kmer (as ordered by the kmer order in the read) that matched at the
expected position. If there was, we extend a "thread" of matching node ids.

Note that we can use the node ids to assess ordering, as we are guaranteed that
an id that is lower than ours is not going to occur after the node in the
partial order. Obviously, this strategy breaks down where we can't meet this
requirement, but it works as long as we approximately can.