greatly improve kmer threading
Remove the "node-to-node" threading. This just picks up lots of nonspecific
mapping. Replace this with an improved positional threading method that relies
on having at least 2 hits on a single node (obviously, the node needs to be >
1bp long, but otherwise this is a small limitation). After the first threading,
link together threads that are < some lenient cutoff (30 nodes) from each
other. 30 nodes is still very small in the scope of the genome! This works
because most of the nonspecific kmers will form only a single hit, and further
won't be ordered. At the end, combine the threads into
single-graphs-per-thread, resolving fragmentation.