use bulk loading techniques
This speeds up index writes by about 100%, which would seem minor but some
indexing jobs take days--- so it's appreciated. Basically the graph is loaded
into the index and its kmers are indexed. The db is closed, reopened with
default options, and then closed again on index object destruction to force
compaction of all SST tables into a single level. This really improves query
performance, although the db will still function if it is not completely
compacted. (One way to get around this issue is to set a bloom filter policy,
which greatly decreases the cost of lookup misses.)