use callbacks to build a general compressed protobuf format
stream.hpp (which should probably be spun out into its own project...) provides
two template functions: stream::write and stream::for_each, which should work
on any protobuf object.

The format is very simple. A compressed file is read. The first 64 bits encode
the number of objects in the file. The next 32 bits encode the following object
size, and so on: [count][[message_size][message] ...].

The write interface calls a callback with a chunk number. The callback is
assumed to use this to decide what chunk of the original object to output. It
could be used as an id into a vector of objects. The callback must return an
object of a particular type to be written to the stream.

The for_each interface calls the callback for each object in the compressed
stream.

Abstracting this interface should make it very easy to