cache intermediate kmer results on disk, and load in one go
To improve the overall parallelism of the kmer indexing process, it may be
necessary to first write to independent files (one per thread, per buffer
size). This completely removes any contention, at least at the application
layer, and hopefully the contention for disk resources will be lower if chunks
of sufficient size can be written.