extend kmer query interface to return actual kmers
The query interface can handle any kmer length shorter than those that have
been indexed. As such, it can be important to determine which actual kmer was
matched, so as to allow exploration of the kmer space of the graph. This
provides a mechanism, and changes the tabular query method of the command-line
API to support it.