alignment rescue, default kmer step == kmer size
Set the default step size (jump between successive kmers pulled from the read)
to be the kmer size, or non-overlapping. If alignment does not succeed, divide
the step size in half and try again. If the step size is already half the kmer
size, decrease the kmer size by 2 and set the step size to half the kmer size.
Do this until kmer_min (12, but possibly it should be lower).

This allows us to use long, non-overlapping kmers for reads where they work,
and then put more work into difficult reads.

An extension would be to specify a score/base threshold that needs to be met,
otherwise we try to realign.