improve mapping performance to 900 reads/s on 1mb test graph
It's important to expand context by a fixed amount. It might be a good idea to
figure out how to dynamically determine if we should expand, as this could win
better. This improves things because sometimes the mapper needs to drop stride
and kmer size to when the initial graph isn't complete enough to map the read.