change edit from pseudo-CIGAR to transformation based
In the new format edits are recorded with full information required to take the
reference and reconstruct the sequence associated with a given mapping to it.

The sequences of SNPs and insertions are recorded. Other cigar-like
"operations" such as soft and hard clipping are supported implicitly as
described in the spec.