☯ huge improvement in alignment performance by undoing bugs
A bug related to taking the "first kmer hit" yielded truncation of the target
regions of the graph used for alignment (threads). In consequence, many more
reads were soft clipped than should have been, sometimes many %. This in turn
triggered the soft clip resolution code on huge numbers of reads, slowing
alignment and frustrating the original objective of the performance tweak.

In addition, a splash was added (vg::Mapper::thread_extension) to avoid the
issue by sampling some number of nodes around the target kmer hits. This was
even more expensive than the first issue.

I uncovered the problem when I noticed that there was a directional bias to the
soft clipping. It was always at the 3' end of the reads!

Tests on my 4-core laptop show this commit running at 1k
reads/CPU/s, which is nearly 2x as fast as previous ones. I'm testing against a
variation graph made from the 1000G data in the test folder (1mb1kgp/) indexed
with "-k 27 -e 7".

☯ Trust the kmers. ☯