remove requirement of VCF input for vg construct
It's sensible to want to make a graph of *just* a reference, against which we
can then align other contigs, reads, and so on to drive forward a more dynamic,
iterative construction. This is also useful for simulating reads from a
reference sequence with vg sim. There was no reason for this limitation, but a
few code tweaks were required to remove it.