SAM/BAM/CRAM: all that's left is a memory leak
My parsing strategy is to feed htslib a string version of the SAM input.
However, this requires creating and destroying readers for every alignment.
Apparently neither sam_close, nor free, nor wind or weather can tear down the
memory that htslib has allocated to the file reader. So there is a big memory
leak in surjection in this commit!