modify a graph with an alignment
An alignment is a graph, and similarly a graph alignment implies a new graph.
With this commit the inclusion of a single alignment in a graph is supported.
However, as the operation is destructive, any paths that are native to the
graph in the region that is modified may no longer be valid.

To rectify this issue, there are two possible approaches. The first is to
ensure alignments may be expressed using a stable named coordinate system
embedded within the graph, for instance a side graph. Then the modification of
the graph needs to not disrupt this annotation, and all alignments will remain
valid.

another approach is to pool the paths which we want to include in one step
together, and modify the graph one node at a time using all the mappings in
each. this may be conceptually simpler, and could be seen as a lower-level
requirement for adding paths to the graph.