resolve traceback issues with byte-sized striped alignments
The word-sized and byte-sized alignment traceback routines drifted out of sync
some time ago. This didn't bode well for byte-size striped alignments, which
tended to develop huge soft clipping. Not good. This incorporates the fix to
gssw and also includes test material to ensure the problem does not recur. Some
tests were updated as the fix also improved the alignment results and
simplified the modified graphs that result from alignment to graph conversion.