Adding support for node orientation to the Index object.
Adding a test for edge orientation in the index.

Orienting the nodes in the graph with a topological sort before aligning to them in the mapper, and flipping the relevant bits of the alignment back afterward.

Edge data is now stored at the key for the end of the edge corresponding to the lower-valued node. Added an overload of parse_edge that just parses the key, so you don't need to make up fake edges to hold the relative orientation.

I'm also changing the flag values on vg find from -t and -f to -s and -e for edges on the start and end of nodes. The option for sequences got bumped from -s to -S.

Making vg find negate node IDs on the other end of reversing edges.