Add support for cycles
This is a fairly substantial revision of some of the internal data
structures and the algorithms that operate on them.

First, modify the VG object to allow storing multiple edges between a
pair of nodes, including self edges. This is accomplished by creating a
new NodeSide type, representing a pair of a node ID and a flag to
differentiate between the start side and the end side. The edge index in
edge_by_id has been renamed to edge_by_sides and modified to index edges
by the sides they connect, instead of the nodes they connect.

Add tests for a variety of sizes and types of cycles, including cycles
between two nodes, self-loop cycles, and cycles which do not connect to
any of the graph's head or tail nodes.

Add tests for a variety of activities with cyclic graphs:
	* GFA orientation flags
	* dot output
	* vg align
	* index alignment storage
	* GCSA2

Change some tests that were previously using hashes to parse vg output
and check semantics.

Revise the way the graph is topologically sorted, so that cycle-closing
edges or components of the graph not connected to any head or tail do
not break the algorithm. The problem of ordering and orienting a graph
"best" in the presence of cycles is NP-hard. In this algorithm, a
heuristic is used, where the heads are first suggested as oriented
"seeds". The algorithm grabs a seed, and for everything to the right of
it, orders and orients every node with no other incoming edges.  Nodes
with multiple incoming edges are added to the collection of seeds.  If
the algorithm runs out of seeds, it starts taking arbitrary nodes, until
every node has been ordered and oriented.

Drop edges that go against the node ordering and orientation when
preparing for GSSW alignment.

Revise the way GCSA2 kmers are generated. Now generates a GCSA2 node for
the forward strand, and one for the reverse strand, and wires them up to
fake reversing edges. Note that this is trickier than just outputting
the same kmer forwards and backwards. A GCSA2 kmer is really a tree; it
can have only one start position, but many end (and after-the-end next
kmer) positions, as long as all the paths it follows spell out the kmer
sequence. So what is one kmer on the forward strand, with two possible
after-the-end positions, can be two kmers on the reverse strand, which
both end at the same place. To make this work, I have modified
_for_each_kmer() and provide for_each_kmer_of_node() to support
iterating over the kmers on both strands that start at a given node.

Add one combination start/end node for GCSA2, instead of distinct start
and end nodes. Because of the way we are now doubling everything for
GCSA2, the resulting GCSA2 graphs appear to have one start node, one end
node, and two of everything else.

Fix the end node in GCSA2 kmers always going to 0:0 instead of whatever
the start node ID actually is.

Revise the start/end node addition procedure to break into disconnected
components that have no heads or tails, and tie them to the start/emd
node at an arbitrary place. This is required for GCSA2 to index the
resulting graph, but necessitated the addition of some connected
components code, like for_each_connected_node().

Revise kmer iteration. Now when one disallows negative-offset kmers,
kmers that were previously simply ignored for being at a negative offset
are now flipped onto the reverse strand, and returned there with a
positive offset. Since this would otherwise result in all kmers spanning
edges being double-counted, kmers that start in one node and end in
another are only yielded from the lower-ID node. Note that this actually
fixes a bug with edge orientation: previously, kmers spanning
start-to-start edges were not being produced, as they had negative
offsets from both nodes' points of view.

Revise the cache used to deduplicate kmers to account for both the
kmer's start position relative to its node and its actual start position
in the graph. Previously, kmers spelling out the same string along
different paths coming into a node were not regognized as distinct when
iterating over kmers for that node.

Replace the use of NodeTraversal to talk about a location on a path
represented by a list<NodeTraversal>. Since cycles can exist in the
graph, the same node can exist in a path multiple times, possibly
traversed in the same direction, and thus this usage of a NodeTraversal
is ambiguous. Instead, a list<NodeTraversal>::iterator is now used,
which allows multiple occurrences of the same NodeTraversal along a path
to be distinguished.

Replace the use of Node* to talk about a location on a Path* for the
same reason. Now a size_t index to the appropriate Mapping along the
Path is used. This makes getting the Node* a bit unweildy, but it is
still an O(1) operation.

Make _for_each_kmer behave properly when called from multiple threads,
with parallel = false. This is necessary to make it properly re-entrant:
multiple MPI threads can now run distinct non-parallel _for_each_kmer
iterations at the same time.

Make GCSA2 kmer generation and indexing ignore the -d option, and remove
the allow_dups flag from the GCSA2 kmer methods. "Duplicate" kmers in
this sense are the kmers that start at a certain position but end at
different positions, and GCSA2 indexing requires that they all be
enumerated in order to have the full set of possible end positions,
which the indexer needs. Previously, passing -d along with -g to vg
kmers would produce output not valid for GCSA2 indexing.