Return edges from edges_for_node only once
Edges that connect the two sides of a node were previously being
returned twice here, which could lead to them being unindexed and
reindexed multiple times if a node needed to be flipped for an
alignment, thus corrupting the graph indexes.

In addition to this change, it would probably be smart to make the
unindexing and reindexing functions throw errors if an edge is unindexed
or reindexed multiple times.