multiple sequence graph aligner
`vg msga` now provides an interface to a basic progressive multiple sequence
aligner. Alignment is still with gssw, and maps one sequence to a graph, then
merges the alignment with the graph. In this way highly divergent haplotypes
may be aligned to each other. A single test is included as proof of principle.