Add multimapping support
Multimapping support is exposed to the user through the -M option to `vg
map`. The user specifies a maximum number of alignments per read, and
that many of the top non-identical alignments for each read (generated
form different seed threads) are output. The top-scoring alignment for
each read is "primary", while the remainder are flagged with a new
`is_secondary` flag in GAM.

Alignments are emitted for a read in descending order of score.

BAM input has been improved: template names have /1 and /2 added to them
to firm unique read names for the reads at each end of the template, and
secondary BAM alignments are only re-aligned if the -K option is passed
to `vg map`.