partly enable large-graph multiple sequence/graph alignment
This is accomplished by using the vg::Mapper rather than VG::align to generate
alignments against the graph. Alignment using the mapper can be banded,
allowing efficiency even for large graphs. Indexing is not as efficient as
alignment for small graphs, but should be much better for large ones up to
hundreds of MB. The current algorithm (in main.cpp) for msga could be improved
in a variety of ways, such as iteratively trying to fit bits of the
graphs/sequences that are being aligned. True MSGA will also require the use of
collections (representing whole graphs) rather than just paths (which imply
walks through the graph with definite order).

This commit surely will break fail tests, so don't use it. But as things are
beginning to work it serves as a checkpoint.