major refactor of path and alignment editing
Working from two sets of path division functions (one for cutting in
reference-relative positions, and the other for offsets in the path itself),
editing alignments when merging becomes much easier.

There are a lot of debugging messages in this commit. It's not meant for use,
just as a local checkpoint.