graph normalization routines
We can generate a normalized graph by removing redundant edges and node
sequences. These are determined using an algorithm that first removes edges
that do not contribute to the path space of the graph, such as edges between
two nodes that have no other edges (unchop); and second finds all sets of nodes
that have identical inbound or outbound node sets and removes redundant
sequence from the start or end of these nodes, pushing it into another node
that is shared (simplify_siblings).  Finally, redundancy that may be introduce
in the last step is removed using unchop.