guarantee that we don't introduce small cycles in normalization
We check that 1) we have a common ancestor between all nodes to be merged, and
2) none of the nodes is an ancestor of any of the others. Failing to make these
two guarantees can cause the introduction of semantically-correct but
otherwise-problematic cycles. The checks are done such that we only proceed
when we can verify that we aren't merging into a cycle, which by default uses a
window of 32 nodes (we could do more but this might be excessive in a
heavily-branching graph). Anyway, the limit can be adjusted with experience.