sync the paths before rebuilding the xg index
If the paths index and the graph's path representation get out of sync we will
fail to build the xg index. I noticed that this is resolved when I write the
graph to file before building the xg index, so I have added a synchronization
step in an appropriate place to try to mitigate the problem.