position rework--- store path directions using positions (resolves #127)
Updates positions to provide the is_reverse flag, and removes this flag from
other places in the schema. The confusing and BAM-derived alignment is_reverse
flag is no more. Mappings can no longer have it either.

Massive commit. 82k diff! I'm checkpointing here because basic editing examples
are working. There are still a ton of bugs here, but they seem to involve
validate and banded alignment. Committing so as to not shoot myself in the
foot.