make a final pass through the path and add any missed edges
I find that in some rare cases the paths are correct but some edges are not
included. It looks like this is happening when long edges jump from preexisting
node ends. However, I haven't verified this, so perhaps it would be better to
leave the warning message about resolving this un-commented. Anyway, this
solidifies the fully-progressive assembly process. I'm passing all the typical
tests and also the HLA genes, although these are too compute-intensive to have
in the automated tests.

The next thing to do here would be performance improvements. That can wait
though while I dig back into mapping.