copy-pasta bug led to normalization failure
It wasn't possible to get the correct "perfect from siblings" because I borked
the instantiation of a NodeSide that was passed to edges_from(NodeSide), making
it behave as if it was for edges_to(...). This made all siblings appear to be
perfect siblings, which in turn destroyed normalization.

Testing on CENX showed the error. That's some good fuzz for an assembler.