prove that unfold->dagify generates a kmer-correct graph
If we first dagify then unfold we can re-introduce cycles. By first dagifying
and then unfolding we  avoid this, so it is important that we can do it
correctly. This test checks a trivial case which is made difficult due to the
generation of two strongly connected components from the input graph by unfold.