Able to convert raptor_term to std::string #220
Managing to be able to talk to VG compile wise in the turtle_parse #220

We have a good start on the turlte parser #220

Nodes written into RDF are parsed out again #220

Sequences attached to nodes parsed out and saved #220

Parsing and adding edges #220

Working on matching the paths into the vg graph. #220 however, mapping not constructed correct.

Working rdf import #220, assumes rdf was generated by vg in the first place.