serial alignment merge with less copying
The parallel merge appears to become disordered with high parallelism.
Additionally, it saves little time due to the huge number of alignment copies
which must be made. In the interest of simplifying things and resolving this
issue, I've moved the merge back to a serial process that builds up a single
alignment.