wire up MEM mapper, resolve MEM resolution problem
If we exhaust the range, we need to step back to the last range, which wasn't
being done with the result being a failure to get anything but the first MEM in
the read.

Also wires MEM mapping into the CLI. For good performance, the neighborhood
expansion parameter to the mapper needs to be increased. I'm finding `-n 5` to
be good. The problem is that we get many soft clips or many subgraphs if we
don't manage to expand the context out the the full subgraph between successive
MEMs that are separated by a mismatch.