hook up MEM mapping in msga
This works. But, it has exposed many bugs due to the change in the way the
alignments work. The MEM approach appears to have a higher intrinsic mapping
rate, and as a result we are getting new kinds of weird structures in the graph
that break normalization routines.

To resolve this I then restructured the remove_node_forwarding_edges code to
use use NodeTraversals to enumerate the edges to forward. I worry that there
was not a bug here, and I haven't run tests to see if this is introducing a
problem, so it may be reverted.

I'm checkpointing here because the rabbit hole of bugs has just gone three
levels deep and it's getting a little too real to continue without rigging up
some climing gear to get out if it gets even more intense.

In rewriting remove_node_forwarding_edges, I generated graphs with many self
loops on nodes.  These are not handled properly in the unfold/dagify unrolling,
it seems; will check. There is a lot of debugging code floating around now
because I'm not sure which bugs have been resolved yet.