incomplete development of an index-driven mapping algorithm
In principle, we should be able to use the series of exact matches provided by
the forward MEM method to determine likely alignment positions. We shouldn't
need to align these segments using GSSW, because we already know they match.
Here, I've attempted to develop a method that povides a single "optimal"
alignment using the forward MEM list for a given read. We convert the exact
matches into alignments. Then, the regions between the forward-MEMs get aligned
in the regions of the graph which are near the exact matches that flank them.
The resulting vector of multi-alignments can be resolved into a single
alignment by passing over it with resolve_banded_multi. This is all nice, but
it does not seem to work well in the presence of high error rates, which
trigger many spurious forward MEM matches. I will come back to this when we
have true MEM functionality in GCSA2.