avoid picking up the paths from the xg index in the mapper
This blocks due to poor design of xg path handling (there is a single threaded
compressed suffix array that's used to store the path names, which should be
changed). In any case, we can avoid getting the paths by specifying the call to
expand_context correctly, which resolves the basic issue of thread blocking.