Name variants with SHA1 hashes if they lack IDs
This is a bit verbose, but I'm worried about the edge case of multiple
variants at the same site, and I don't want to encode the whole variant
allele set in the allele names. Maybe I can make the names shorter if I
can list the alleles because they are short?