remove normalization from MSGA
Strict graph normalization is not required to make an assembly, and a single
round may not be enough to achieve stability. To make this explicit, I'm
un-setting the normalization process that ran at every step of MSGA. Tests are
updated  to reflect this change, sometimes by the use of vg mod -n to achieve
the same result.