SMEM clustering rather than brute force expansion + subgraphing
We lift the concept of thread extension from the kmer alignment algorithm into
the SMEM one and cluster node hits which are under a given id space distance
(the --thread-ex).

This is meant to avoid the situation in which many small spurious hits generate
extremely large graph agglomerations in dense graphs formed by assemblies.

The second resolution for this is forthcoming, and involves avoiding alignmnet
to graphs over a given configurable size.

This reintroduces a slight id-space dependency in the mapper, but the effect
should be mild. Context expansion based on graph distances is still applied to
the subgraph that results from each cluster.