change to read length multiple based alignment target graph cutoff
If the target graph in the mapper is 100X as long as the read, don't try to
align. The index should be more precise than that, and if it's this long we run
a risk of generating a really large alignment target if we also need to unroll
some components.