initial implementation of path-only indexing
All aspects of vg which are important for getting kmers and making the GCSA2
index have been touched to optionally ignore paths in the graph that do not lie
in any path. This is stil not quite working, because it is ignoring the crucial
head and tail nodes for GCSA2. There may be other problems as of this commit.