semi-failed attempt to resolve problems with path-only indexes
Indexing path-only with GCSA2 is still not working, although I did resolve what
seemed to be the basic problem. I have switched MSGA back to index the entire
graph with recombinations by default.

I am inclined to use the approach I have worked out as it is natural to extend
to threads in gPBWT. However, it might be simpler to just pull the kmers out of
the paths. It will be much harder to do this well when we have 10k+ haplotypes
in the graph.