require specification of rocksdb index
This normalizes use of the xg, gcsa2, and rocksdb indexes so that each one must
be explicitly given. Before we were auto-deciding the rocksdb index in certain
situations, which made it hard to ensure that the rocksdb code path was
completely separate from the others. I also found it confusing, and had been
meaning to do this for a while.