Add paired-end read consistency constraints
Now paired-end reads will be looked at jointly, and the primary
alignments for each read chosen so as to produce a pair of alignments
that are close together in the graph. All "consistent" pairs of
alignments are considered, and the consistent pair with the highest
total alignment score is made primary.

We look for consustent pairs among the mappings of each end, and among
the mappings we get rescuing each end of the read of each of the
mappings of the other end.

The -M multimapping option can affect this by controlling he number of
individual read alignments used. If the best pair is really the
third-best mapping of read 1 and the third-best mapping of read 2
(because the top 2 mappings of each read have no good consistent
mappings for the other read), that optimal pair won't be found unless
you are using at least -M 3. Also, because we consider all pairs of
mappings, the mapper is now O(n^2) in the -M parameter.

If a consistent pair cannot be found, the primary mappings will be the
two highest-scoring mappings, regardless of consistency.

Don't try to anchor on unmapped reads

Make sure that reverse rescued alignments look reverse