allow superbubble-based organization of dot output
If we have a DAG with large numbers of alignments, we need to tell dot which
nodes should be ranked together, or it will tend to randomly shift them around,
increasing the visual complexity of the graph.

We now can do this with `vg view -dB`. This is optional because superbubbles
don't work unless we have a DAG.