implements alignment patching for MEM-threaded alignments
The unaligned fragments of read clusters are now patched into the reference
graph by way of local alignment to the fragment of the graph that the rest of
the alignment implies is nearby.

Read scores are not yet implemented, and this commit carries in a huge number
of debugging flags that will need to be removed.

The alignment process has only been tested on a trivial example.