MEM threader (#433)
* remove long-unused commented-out code

* paper: add some references

* updates GCSA2, fixes mapping test with incorrect command line arg

* msga: build graph in order of sequence input

* msga: nonce string names when passed raw sequences

This allows us to have duplicate sequences without the
inclusion failing.

* msga: change nonce strategy to stabilize hash-based names

* resolve documentation to fix #415

* adds -g flag, useful for profiling

* use path position feature to build a SMEM clustering solution

* partial implementation of SMEM-threaded alignment

* allow traceback to be forced to begin at match on end of ref and read

This is essential for using DP in the gaps in the
SMEM-threaded alignment.

* force traceback to start from a particular node

In preparation for filling in the gaps in SMEM-threaded
alignment in the mapper.

* don't try to nail down alignment traceback begin

* implements alignment patching for MEM-threaded alignments

The unaligned fragments of read clusters are now patched into the reference
graph by way of local alignment to the fragment of the graph that the rest of
the alignment implies is nearby.

Read scores are not yet implemented, and this commit carries in a huge number
of debugging flags that will need to be removed.

The alignment process has only been tested on a trivial example.

* comment out a lot of debugging flags and translate cut node mappings

* fixes up trimmed node position translation logic

* slow version of mem threading

The implementation has exposed us to some new issues.

This is to mark the "correct" version in which we don't assume a partial order
sort in the patching of the SMEM thread.

* cache nodes to avoid repeated full node lookups from xg

Improve the pattern in the Sampler, and then move to using the same code (now
in position.cpp) to do so in both the Mapper and Sampler.

* remove dead/commented out code for next position in mapper

* complete refactor of match walker interface

* expose MEM threader on command line (for testing)

* bailout in expand_context when we have nowhere to extend

This saves work and allows for full subgraph exploration.

* avoid calling disjoint_subgraphs in Mapper::patch_alignment

* cleans up a number of problems with mem threader

Including outright failing to build the mapping correctly.

* only run the threading code for up to the number of needed multimaps

* remove useless sort

* only unfold and dagify if we don't already have a non-inverting DAG

* implement a faster acyclicity check

* updates xg

* add break function to dfs to fix acyclic check

* avoid aligner race in banded alignment and fix is_acyclic check

Resolves #448

* determine thread count as default for msga aligner_threads

* use the get_aligner function

* mapper changes introduced a change in the truth for translation test

* for some reason there are now 27 subjects in tiny.ttl

* use correct aligner, improve threading/clustering interface, remove debugging code

* on travis, there are 25 subjects, but locally there are 27?

* remove debugging output

* add test for MEM threading

* respect per-Mapper thread count

When banding, we may use many threads in one Mapper. This ensures we respect
the limit. Previous commits got us most of the way.